relocate recentlyOffline state
