+++ packages/app-staking/src/Overview/index.tsx
@@ -19,7 +19,7 @@ const ZERO = new Balance(0);
 
 export default class Overview extends React.PureComponent<Props> {
   render () {
-    const { balances, balanceArray, intentions, nominators, validators } = this.props;
+    const { balances, balanceArray, intentions, nominators, recentlyOffline, validators } = this.props;
     const intentionsSorted = this.sortByBalance(
       intentions.filter((address) =>
         !validators.includes(address)
@@ -40,6 +40,7 @@ export default class Overview extends React.PureComponent<Props> {
           current={validatorsSorted}
           next={intentionsSorted}
           nominators={nominators}
+          recentlyOffline={recentlyOffline}
         />
       </div>
     );
